Move tab component into tabs subfolder
